feat(BUX-166): method for getting merkle paths from arc removed
