+++ model_bump_test.go
@@ -929,120 +929,3 @@ func TestBUMPModel_MerkleProofToBUMP(t *testing.T) {
 		assert.Equal(t, BUMP{BlockHeight: blockHeight}, actualBUMP)
 	})
 }
-
-// TestBUMPModel_MerklePathToBUMP will test the method MerklePathToBUMP()
-func TestBUMPModel_MerklePathToBUMP(t *testing.T) {
-	t.Parallel()
-
-	t.Run("Valid Merkle Path #1", func(t *testing.T) {
-		// given
-		blockHeight := uint64(0)
-		mp := bc.MerklePath{
-			Index: 1,
-			Path:  []string{"txId", "node0", "node1", "node2", "node3"},
-		}
-		expectedBUMP := BUMP{
-			BlockHeight: blockHeight,
-			Path: [][]BUMPLeaf{
-				{
-					{Offset: 0, Hash: "node0"},
-					{Offset: 1, Hash: "txId", TxID: true},
-				},
-				{
-					{Offset: 1, Hash: "node1"},
-				},
-				{
-					{Offset: 1, Hash: "node2"},
-				},
-				{
-					{Offset: 1, Hash: "node3"},
-				},
-			},
-		}
-
-		// when
-		actualBUMP := MerklePathToBUMP(&mp, blockHeight)
-
-		// then
-		assert.Equal(t, expectedBUMP, actualBUMP)
-	})
-
-	t.Run("Valid Merkle Path #2", func(t *testing.T) {
-		// given
-		blockHeight := uint64(0)
-		mp := bc.MerklePath{
-			Index: 14,
-			Path:  []string{"txId", "node0", "node1", "node2", "node3", "node4"},
-		}
-		expectedBUMP := BUMP{
-			BlockHeight: blockHeight,
-			Path: [][]BUMPLeaf{
-				{
-					{Offset: 14, Hash: "txId", TxID: true},
-					{Offset: 15, Hash: "node0"},
-				},
-				{
-					{Offset: 6, Hash: "node1"},
-				},
-				{
-					{Offset: 2, Hash: "node2"},
-				},
-				{
-					{Offset: 0, Hash: "node3"},
-				},
-				{
-					{Offset: 1, Hash: "node4"},
-				},
-			},
-		}
-
-		// when
-		actualBUMP := MerklePathToBUMP(&mp, blockHeight)
-
-		// then
-		assert.Equal(t, expectedBUMP, actualBUMP)
-	})
-
-	t.Run("Valid Merkle Path #3 - with *", func(t *testing.T) {
-		// given
-		blockHeight := uint64(0)
-		mp := bc.MerklePath{
-			Index: 14,
-			Path:  []string{"txId", "*", "node1", "node2", "node3", "node4"},
-		}
-		expectedBUMP := BUMP{
-			BlockHeight: blockHeight,
-			Path: [][]BUMPLeaf{
-				{
-					{Offset: 14, Hash: "txId", TxID: true},
-					{Offset: 15, Duplicate: true},
-				},
-				{
-					{Offset: 6, Hash: "node1"},
-				},
-				{
-					{Offset: 2, Hash: "node2"},
-				},
-				{
-					{Offset: 0, Hash: "node3"},
-				},
-				{
-					{Offset: 1, Hash: "node4"},
-				},
-			},
-		}
-
-		// when
-		actualBUMP := MerklePathToBUMP(&mp, blockHeight)
-
-		// then
-		assert.Equal(t, expectedBUMP, actualBUMP)
-	})
-
-	t.Run("Empty Merkle Path", func(t *testing.T) {
-		blockHeight := uint64(0)
-		mp := bc.MerklePath{}
-		actualBUMP := MerklePathToBUMP(&mp, blockHeight)
-		assert.Equal(t, BUMP{BlockHeight: blockHeight}, actualBUMP)
-	})
-}
